Rename to register2 for ease of merge
