+++ user/src/main/java/com/example/user/presentation/register2/RegisterViewEvent.kt
@@ -1,0 +1,4 @@
+package com.example.user.presentation.register2
+
+sealed class RegisterViewEvent() {
+}
