+++ user/src/main/java/com/example/user/presentation/register2/RegisterViewState.kt
@@ -1,4 +1,4 @@
-package com.example.user.presentation.register
+package com.example.user.presentation.register2
 
 import java.io.Serializable
