+++ user/src/main/res/layout/fragment_register2.xml
